refactor: move the theme folder
