+++ app/src/main/java/id/fishku/fisherseller/presentation/ui/dashboardv2/components/WeatherAndTideCard.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import id.fishku.fisherseller.R
-import id.fishku.fisherseller.presentation.ui.dashboardv2.theme.fonts
+import id.fishku.fisherseller.compose.theme.fonts
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
