+++ app/src/main/java/id/fishku/fisherseller/presentation/ui/home/component/BottomSheetMore.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 import id.fishku.fisherseller.R
-import id.fishku.fisherseller.presentation.ui.dashboardv2.theme.fonts
+import id.fishku.fisherseller.compose.theme.fonts
 import id.fishku.fishersellercore.model.MenuModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
